refactor(tui): render.py を text_format.py にリネームする
中身は表示幅を意識した文字列の整形で、Renderable を組み立てる app_render.py
とは層が違う。render という名前だと紛らわしいため改める。
