+++ src/redi/tui/text_format.py
